refactor: move add command to subpackage
