+++ cmd/add/add.go
@@ -1,9 +1,7 @@
-package cmd
+package add
 
 import (
 	"fmt"
-	"net/http"
-	"net/url"
 	"strings"
 	"time"
 
@@ -11,7 +9,6 @@ import (
 	"github.com/nikcorg/tldr-cli/config"
 
 	"github.com/urfave/cli"
-	"golang.org/x/net/html"
 )
 
 // Add is the Add command
@@ -54,6 +51,7 @@ func (cmd Add) Configure(c *config.Config) cli.Command {
 				if err != nil {
 					panic(err)
 				}
+
 				title = fetchedTitle
 			}
 
@@ -71,58 +69,22 @@ func (cmd Add) Run(url string, title string) {
 		Entries: []archive.Entry{},
 	}
 
-	fmt.Printf("new archive %+v\n", a)
-
 	if archive.Exists(filename) {
 		a = archive.Load(filename)
 	}
 
-	a.Entries = append(a.Entries, archive.Entry{
+	entry := archive.Entry{
 		Title:   title,
 		URL:     url,
 		Unread:  cmd.flags.Unread,
 		Related: toSimpleEntrySlice(cmd.flags.Related),
 		Source:  cmd.flags.Source,
-	})
+	}
+	a.Entries = append(a.Entries, entry)
 
 	archive.Save(filename, a)
-}
 
-func retrieveTitle(href string) (string, error) {
-	resp, err := http.Get(href)
-	if err != nil {
-		panic(fmt.Errorf("unable to fetch title from %s: %+v", href, err))
-	}
-
-	doc, err := html.Parse(resp.Body)
-	if err != nil {
-		panic(fmt.Errorf("unable to parse the html from %s: %v", href, err))
-	}
-
-	title := titleNode(doc)
-
-	if title == "" {
-		return "", fmt.Errorf("unable to retrieve title for %s", href)
-	}
-
-	return title, nil
-}
-
-func titleNode(n *html.Node) string {
-	if n == nil {
-		return ""
-	} else if n.Type == html.ElementNode && n.Data == "title" {
-		return n.FirstChild.Data
-	}
-
-	for c := n.FirstChild; c != nil; c = c.NextSibling {
-		tit := titleNode(c)
-		if tit != "" {
-			return tit
-		}
-	}
-
-	return ""
+	fmt.Printf("Title: %s\n  URL: %s\n", entry.Title, entry.URL)
 }
 
 func toSimpleEntrySlice(xs []string) []archive.RelatedEntry {
@@ -131,12 +93,4 @@ func toSimpleEntrySlice(xs []string) []archive.RelatedEntry {
 		es[i] = archive.SimpleEntry{URL: url}
 	}
 	return es
-}
-
-func assertIsHref(href string) string {
-	if v, e := url.Parse(href); e != nil || !v.IsAbs() {
-		panic(fmt.Errorf("Trying to add non-absolute URL: %s", href))
-	} else {
-		return v.String()
-	}
 }
